requirements.txt
Django==5.0
requests
aiogram
psycopg2-binary
djangorestframework
django-filter